Renamed GiantRats.js file
#8
