+++ src/Monsters/Tables/Level1.js
@@ -9,7 +9,7 @@ import Monsters from "../Races/Monsters";
 import {CaveSquig} from "../Races/Monsters/CaveSquigs";
 import {Centaur} from "../Races/Monsters/Centaurs";
 import {GiantBat} from "../Races/Monsters/GiantBat";
-import {GiantRat} from "../Races/Monsters/GiantRat";
+import {GiantRat} from "../Races/Monsters/GiantRats";
 import {GiantSpider} from "../Races/Monsters/GiantSpider";
 import {Minotaur} from "../Races/Monsters/Minotaurs";
 import {Ogre} from "../Races/Monsters/Ogre";
